requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.4
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-filter==24.3
django-redis==5.4.0
django-rest-framework==0.1.0
django-safedelete==1.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
isodate==0.7.2
kombu==5.4.2
lxml==5.3.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
zeep==4.3.1
